refactor: extract QuestionRunnerEngine and refactor runners
