+++ app/library/[topicId]/TopicContent.tsx
@@ -1,8 +1,9 @@
 import { getQuestions } from '@/api/trainer.api';
 import { getWidgetsByTopic } from '@/api/widget.api';
-import QuestionsRunner from '@/components/library/widget/runners/default/QuestionRunner';
+import DefaultRunner from '@/components/library/widget/runners/default/DefaultRunner';
+import { SliderRunner } from '@/components/library/widget/runners/slider/SliderRunner';
 import WidgetList from '@/components/WidgetList';
-import { toWidgetFilter } from '@/types/widget';
+import { toWidgetFilter, WidgetType } from '@/types/widget';
 
 type TopicContentProperties = {
   topicId: string;
@@ -18,5 +19,10 @@ export default async function TopicContent({ topicId, widgetType }: TopicContent
   }
 
   const questions = await getQuestions(topicId, selectedFilter);
-  return <QuestionsRunner questions={questions} />;
+
+  if (widgetType === WidgetType.FlipCard) {
+    return <SliderRunner questions={questions} />;
+  }
+
+  return <DefaultRunner questions={questions} />;
 }
